Sign package content in a larger HFS volume
The previous approach of simply resizing the dmg image led to the creation
of a corrupted package.

User was prompted with the message "There may be a problem with the disk.
Are you sure you want to open it ?"

Running "fsck_hfs image.dmg" also confirmed the problem:

```
$ hdiutil mount ./Slicer-4.10.0-macosx-amd64.dmg
[...]
Agree Y/N? Y

There may be a problem with this disk image.  Are you sure you want to open it?
[Opening this disk image may make your computer less secure or cause other problems.]
Attach? (Y/N) Y

$ fsck_hfs /dev/disk3s1
** /dev/rdisk3s1 (NO WRITE)
   Executing fsck_hfs (version hfs-305.10.1).
** Checking non-journaled HFS Plus Volume.
   Invalid number of allocation blocks
(4294967295, 0)
** The volume   could not be verified completely.
```